+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -8,7 +8,7 @@ CHANGELOG
  * deprecated FormPerformanceTestCase and FormIntegrationTestCase in the Symfony\Component\Form\Tests namespace and moved them to the Symfony\Component\Form\Test namespace
  * deprecated TypeTestCase in the Symfony\Component\Form\Tests\Extension\Core\Type namespace and moved it to the Symfony\Component\Form\Test namespace
  * changed FormRenderer::humanize() to humanize also camel cased field name
- * added FormProcessorInterface and FormInterface::process()
+ * added RequestHandlerInterface and FormInterface::handleRequest()
  * deprecated passing a Request instance to FormInterface::bind()
  * added options "method" and "action" to FormType
  * deprecated option "virtual" in favor "inherit_data"